fix storybook issue
